refactor(frontend): migrate nodes tab to a route + extract useSourceView (#3962 5.4 PR4)
Moves the `nodes` tab (App's default/index tab) from the
`activeTab === 'nodes' && …` render gate to its own `<Route path="nodes">`,
and extracts the shared node/traceroute/map orchestration it depends on
(processedNodes, marker refs, traceroute path rendering, and the
favorite/delete/purge handlers) into a new `useSourceView` hook
(src/hooks/useSourceView.ts) per task54_spec.md §3/§7.

App.tsx destructures the hook's return into the same local names its other
not-yet-migrated consumers (the inline messages block, PurgeDataModal,
NodePopup) already close over, so those callers are unchanged. The
module-level pending-favorite/ignored/hide-from-map toggle singletons move
from App.tsx into src/utils/pendingToggles.ts so both App's poll
reconciliation and the hook's toggleFavorite/toggleFavoriteLock can import
them without a circular dependency.

NodesTab itself is untouched — same props, same shapes.
